refactor: moved Observable trait implementation for atom & atom_undo & reaction to their respective files
